Remove features/pricing, move queries & mutations
