--- src/features/pricing/styles.tsx
@@ -1,47 +1,0 @@
-import styled from 'react-emotion'
-
-export const HeaderContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-top: 0em;
-`
-
-export const SubHeaderContainer = styled.div`
-  margin-top: 0.8em;
-  margin-bottom: 0.8em;
-  justify-content: space-between;
-  width: 100%;
-`
-
-export const InputContainer = styled.div`
-  margin-top: 2em;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-`
-
-export const PropContainer = styled.div`
-  margin-top: 0.6em;
-  justify-content: space-between;
-  width: 100%;
-`
-
-export const ListContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-`
-
-export const ItemPage = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  max-width: 80%;
-  margin: 0 auto 50px;
-`
